Add strcasestr to windows.
git-svn-id: file:///home/jj/hercules.svn/trunk@6232 956126f8-22a0-4046-8f4a-272fa8102e63

+++ hao.c
@@ -169,7 +169,7 @@ static void hao_cpstrp(char *dest, char *src)
   for(i = 0; src[i] == ' '; i++);
   strncpy(dest, &src[i], HAO_WKLEN);
   dest[HAO_WKLEN-1] = 0;
-  for(i = strlen(dest); i && dest[i - 1] == ' '; i--);
+  for(i = (int)strlen(dest); i && dest[i - 1] == ' '; i--);
   dest[i] = 0;
 }
